Initial admin panel and webapi features added, more to follow.
Today I  was able to delve into the /reviews/admin.py, and I was able
to set up an initial representation of the relationships within the
admin panel. It’s full functionality is still waiting on the web API
connections to be completed, but that will be referenced later.

Next, I was able to jump into the /reviews/views.py, and begin to
really research the inner workings of ReviewBoard. I was able to
significantly improve my understandings of both the Django view
framework and how it relates to this project, so it’s looking quite
promising for now.

This new perspective allowed me to finally delve head first into the
monster that is the /webapi/resources.py file, and I was able to do
some initial progress adding the dependencies into the whole mix.
Specifically, I was able to contribute to the _set_draft_field_data
function within the ReviewRequestDraftResource class, adding in the
dependencies.