Dependency WIP Submission Commit
This commit has the task to a point where a user is able to add
dependencies of drafts and published requests to a review by
means of the Admin panel. At this moment, the blocking reverse
dependency feature is in an early phase, and the front-end means
of dependency interactions are in planning stages.